vector-database-practice-project/data-collection-scripts/Guardian/requirements.txt
requests
pandas
tqdm
python-dotenv
